+++ packages/components/src/ObjectViewer/JSONLike/JSONLike.component.js
@@ -311,6 +311,7 @@ function UntranslatedComplexItem(props) {
 				{info.keys.map((key, i) => {
 					const childId = id && `${id}-${key}`;
 					return (
+						// eslint-disable-next-line @typescript-eslint/no-use-before-define
 						<Item
 							{...props}
 							key={childId || i}
@@ -327,7 +328,19 @@ function UntranslatedComplexItem(props) {
 			</ul>
 		);
 	}
-
+	const childCount = (
+		<sup
+			key="badge"
+			className={`${theme.badge} badge`}
+			aria-label={t('TC_OBJECT_VIEWER_NB_CHILD', {
+				defaultValue: 'Contains {{count}} child object',
+				defaultValue_plural: 'Contains {{count}} child objects',
+				count: info.length,
+			})}
+		>
+			{decoratedLength}
+		</sup>
+	);
 	return (
 		<LineItem
 			{...props}
@@ -351,23 +364,18 @@ function UntranslatedComplexItem(props) {
 				/>
 			}
 			badge={
-				<TooltipTrigger
-					key="badge-tooltip"
-					className="offset"
-					label={getDataAbstract(data)}
-					tooltipPlacement="right"
-				>
-					<sup
-						key="badge"
-						className={`${theme.badge} badge`}
-						aria-label={t('TC_OBJECT_VIEWER_NB_CHILD', {
-							defaultValue: 'Contains {{count}} child',
-							count: info.length,
-						})}
+				props.hideTooltip ? (
+					childCount
+				) : (
+					<TooltipTrigger
+						key="badge-tooltip"
+						className="offset"
+						label={getDataAbstract(data)}
+						tooltipPlacement="right"
 					>
-						{decoratedLength}
-					</sup>
-				</TooltipTrigger>
+						{childCount}
+					</TooltipTrigger>
+				)
 			}
 			type={props.showType ? info.type : null}
 		>
@@ -400,6 +408,7 @@ UntranslatedComplexItem.propTypes = {
 	onToggle: PropTypes.func.isRequired,
 	opened: PropTypes.arrayOf(PropTypes.string).isRequired,
 	showType: PropTypes.bool,
+	hideTooltip: PropTypes.bool,
 	t: PropTypes.func,
 };
 
@@ -533,7 +542,9 @@ export function JSONLike({ onSubmit, className, style, ...props }) {
 	if (rootIsObject) {
 		if (props.rootLabel) {
 			labelId = (props.id && `${props.id}-label`) || 'tc-object-viewer-label';
-			label = (
+			label = props.hideTooltip ? (
+				<div className={theme['root-label-overflow']}>{props.rootLabel}</div>
+			) : (
 				<TooltipTrigger key="label" label={props.rootLabel} tooltipPlacement="right">
 					<div className={theme['root-label-overflow']}>{props.rootLabel}</div>
 				</TooltipTrigger>
@@ -542,6 +553,7 @@ export function JSONLike({ onSubmit, className, style, ...props }) {
 	}
 	const containerProps = {
 		id: props.id && `${props.id}-container`,
+		'data-testid': props.id && `${props.id}-container`,
 		className: classNames('tc-object-viewer', theme.container, className),
 		style,
 	};
@@ -591,6 +603,7 @@ JSONLike.propTypes = {
 	style: PropTypes.object,
 	rootLabel: PropTypes.string,
 	tupleLabel: PropTypes.string,
+	hideTooltip: PropTypes.bool,
 };
 
 export default withTreeGesture(JSONLike);