+++ packages/components/src/SubHeaderBar/SubHeaderBar.component.js
@@ -115,7 +115,7 @@ function SubHeaderBar({
 						<Renderer.Action
 							id="backArrow"
 							onClick={onGoBack}
-							label={t('BACK_ARROW_TOOLTIP', { defaultValue: 'Go Back' })}
+							label={t('BACK_ARROW_TOOLTIP', { defaultValue: 'Go back' })}
 							icon="talend-arrow-left"
 							bsStyle="link"
 							data-feature={goBackDataFeature}