+++ packages/faceted-search/src/components/Badges/BadgeCheckboxes/BadgeCheckboxesForm.component.test.js
@@ -1,7 +1,8 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
-import { mount } from 'enzyme';
-import { act } from 'react-dom/test-utils';
+
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+
 import { BadgeCheckboxesForm } from './BadgeCheckboxesForm.component';
 import getDefaultT from '../../../translate';
 
@@ -35,12 +36,18 @@ describe('BadgeCheckboxesForm', () => {
 			t,
 		};
 		// When
-		const wrapper = mount(<BadgeCheckboxesForm {...props} />);
+		render(<BadgeCheckboxesForm {...props} />);
 		// Then
-		expect(wrapper.find('.checkbox > label > span').at(0).text()).toBe('Checkbox One');
-		expect(wrapper.find('.checkbox > label > span').at(1).text()).toBe('Checkbox Two');
-		expect(wrapper.find('.checkbox > label > span').at(2).text()).toBe('Checkbox Three');
-		expect(wrapper.find('input[type="checkbox"]')).toHaveLength(3);
+		expect(
+			screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one').nextSibling,
+		).toHaveTextContent('Checkbox One');
+		expect(
+			screen.getByTestId('badge-checkbox-form-checkbox-checkbox-two').nextSibling,
+		).toHaveTextContent('Checkbox Two');
+		expect(
+			screen.getByTestId('badge-checkbox-form-checkbox-checkbox-three').nextSibling,
+		).toHaveTextContent('Checkbox Three');
+		expect(screen.getAllByRole('checkbox')).toHaveLength(3);
 	});
 	it('should trigger on change callback when checkbox generated from checkbox values are clicked', () => {
 		// Given
@@ -55,16 +62,40 @@ describe('BadgeCheckboxesForm', () => {
 			t,
 		};
 		// When
-		const wrapper = mount(<BadgeCheckboxesForm {...props} />);
-		expect(wrapper.find('input[type="checkbox"]').at(0).prop('checked')).toBe(false);
-		wrapper
-			.find('input[type="checkbox"]')
-			.at(0)
-			.simulate('change', { target: { checked: true } });
+		render(<BadgeCheckboxesForm {...props} />);
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one')).not.toBeChecked();
+		userEvent.click(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one'));
 		// Then
 		expect(onChange).toHaveBeenCalledTimes(1);
 		expect(onChange.mock.calls[0][1]).toEqual([
 			{ checked: true, id: 'checkbox-one', label: 'Checkbox One' },
+		]);
+	});
+	it('should trigger on change callback when all checkbox is checked', () => {
+		// Given
+		const onChange = jest.fn();
+		const props = {
+			checkboxValues,
+			id: 'myId',
+			onChange,
+			onSubmit: jest.fn(),
+			value: [],
+			feature: 'Connection type',
+			allSelector: true,
+			t,
+		};
+		// When
+		render(<BadgeCheckboxesForm {...props} />);
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one')).not.toBeChecked();
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-two')).not.toBeChecked();
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-three')).not.toBeChecked();
+		userEvent.click(screen.getByTestId('badge-checkbox-form-checkbox-selectAll'));
+		// Then
+		expect(onChange).toHaveBeenCalledTimes(1);
+		expect(onChange.mock.calls[0][1]).toEqual([
+			{ checked: true, id: 'checkbox-one', label: 'Checkbox One' },
+			{ checked: true, id: 'checkbox-two', label: 'Checkbox Two' },
+			{ checked: true, id: 'checkbox-three', label: 'Checkbox Three' },
 		]);
 	});
 	it('should display checkbox one checked', () => {
@@ -85,9 +116,42 @@ describe('BadgeCheckboxesForm', () => {
 			t,
 		};
 		// When
-		const wrapper = mount(<BadgeCheckboxesForm {...props} />);
+		render(<BadgeCheckboxesForm {...props} />);
 		// Then
-		expect(wrapper.find('input[id="checkbox-one-checkbox"]').prop('checked')).toBe(true);
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one')).toBeChecked();
+	});
+	it('should display all checkbox checked', () => {
+		// Given
+		const props = {
+			checkboxValues,
+			id: 'myId',
+			onChange: jest.fn(),
+			onSubmit: jest.fn(),
+			value: [
+				{
+					checked: true,
+					id: 'checkbox-one',
+					label: 'Checkbox One',
+				},
+				{
+					checked: true,
+					id: 'checkbox-two',
+					label: 'Checkbox Two',
+				},
+				{
+					checked: true,
+					id: 'checkbox-three',
+					label: 'Checkbox Three',
+				},
+			],
+			allSelector: true,
+			feature: 'Connection type',
+			t,
+		};
+		// When
+		render(<BadgeCheckboxesForm {...props} />);
+		// Then
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-selectAll')).toBeChecked();
 	});
 	it('should filter the displayed checkbox using the filter bar', () => {
 		// Given
@@ -101,19 +165,27 @@ describe('BadgeCheckboxesForm', () => {
 			t,
 		};
 		// When
-		const wrapper = mount(<BadgeCheckboxesForm {...props} />);
+		render(<BadgeCheckboxesForm {...props} />);
 
-		expect(wrapper.find('input[id="checkbox-one-checkbox"]')).toHaveLength(1);
-		expect(wrapper.find('input[id="checkbox-two-checkbox"]')).toHaveLength(1);
-		expect(wrapper.find('input[id="checkbox-three-checkbox"]')).toHaveLength(1);
-		act(() => {
-			wrapper.find('input[type="search"]').simulate('change', { target: { value: 'one' } });
-		});
-		wrapper.update();
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one')).toBeVisible();
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-two')).toBeVisible();
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-three')).toBeVisible();
+
+		userEvent.type(
+			screen.getByRole('searchbox', {
+				name: /find a column/i,
+			}),
+			'One',
+		);
+
 		// Then
-		expect(wrapper.find('input[id="checkbox-one-checkbox"]')).toHaveLength(1);
-		expect(wrapper.find('input[id="checkbox-two-checkbox"]')).toHaveLength(0);
-		expect(wrapper.find('input[id="checkbox-three-checkbox"]')).toHaveLength(0);
+		expect(screen.getByTestId('badge-checkbox-form-checkbox-checkbox-one')).toBeVisible();
+		expect(
+			screen.queryByTestId('badge-checkbox-form-checkbox-checkbox-two'),
+		).not.toBeInTheDocument();
+		expect(
+			screen.queryByTestId('badge-checkbox-form-checkbox-checkbox-three'),
+		).not.toBeInTheDocument();
 	});
 
 	it('should call the submit callback', () => {
@@ -129,25 +201,13 @@ describe('BadgeCheckboxesForm', () => {
 			t,
 		};
 		// When
-		const wrapper = mount(<BadgeCheckboxesForm {...props} />);
-		wrapper.find('button[type="submit"]').simulate('submit');
+		render(<BadgeCheckboxesForm {...props} />);
+		userEvent.click(
+			screen.getByRole('button', {
+				name: /apply/i,
+			}),
+		);
 		// Then
 		expect(onSubmit).toHaveBeenCalledTimes(1);
-	});
-	it('should display a button "Apply"', () => {
-		// Give
-		const props = {
-			checkboxValues,
-			id: 'myId',
-			onChange: jest.fn(),
-			onSubmit: jest.fn(),
-			value: [],
-			feature: 'Connection type',
-			t,
-		};
-		// When
-		const wrapper = mount(<BadgeCheckboxesForm {...props} />);
-		// Then
-		expect(wrapper.find('button[type="submit"]').text()).toBe('Apply');
 	});
 });